[개발] ErrorBoundary (#23)
- 에러바운더리 설정
- 에러 날려주기 위해서 상태값 설정
- 뭔가 좀...어거지로 한 거 같기도 합니다..
