+++ package.json
@@ -22,6 +22,7 @@
     "react": "^18.2.0",
     "react-daum-postcode": "^3.1.3",
     "react-dom": "^18.2.0",
+    "react-error-boundary": "^4.0.12",
     "react-hook-form": "^7.45.2",
     "react-redux": "^8.1.2",
     "react-responsive": "^9.0.2",
